+++ src/backend/joanie/tests/demo/test_commands_create_dev_demo.py
@@ -33,6 +33,7 @@ class CreateDevDemoTestCase(TestCase):
         nb_product_credential += (
             1  # create_product_credential_purchased_with_certificate
         )
+        nb_product_credential += 5  # one order of each state
 
         nb_product = nb_product_credential + nb_product_certificate
         nb_product += 1  # Become a certified botanist gradeo